+++ foam-cutter/Machine/State.cs
@@ -2,7 +2,7 @@ namespace FoamCutter.Machine;
 
 public class State
 {
-	private const double COMPARISON_PRECISION = 0.001;
+	private const decimal COMPARISON_PRECISION = 0.001m;
 
 	private readonly Config _config;
 
@@ -11,11 +11,11 @@ public class State
 		_config = machineConfig;
 	}
 
-	public float X { get; set; }
+	public decimal X { get; set; }
 
-	public float Y { get; set; }
+	public decimal Y { get; set; }
 
-	public float Z { get; set; }
+	public decimal Z { get; set; }
 
 	public CoordinateMode MovementMode { get; set; }
 
@@ -25,11 +25,11 @@ public class State
 
 	public bool CuttingOrScoring => Cutting || Scoring;
 
-	public bool AtCoordinates(float x, float y) => CoordinateEquals(X, x) && CoordinateEquals(Y, y);
+	public bool AtCoordinates(decimal x, decimal y) => CoordinateEquals(X, x) && CoordinateEquals(Y, y);
 
-	public bool AtCoordinates(float x, float y, float z) => CoordinateEquals(X, x) && CoordinateEquals(Y, y) && CoordinateEquals(Z, z);
+	public bool AtCoordinates(decimal x, decimal y, decimal z) => CoordinateEquals(X, x) && CoordinateEquals(Y, y) && CoordinateEquals(Z, z);
 
-	private static bool CoordinateEquals(float a, float b) => Math.Abs(a - b) <= COMPARISON_PRECISION;
+	private static bool CoordinateEquals(decimal a, decimal b) => Math.Abs(a - b) <= COMPARISON_PRECISION;
 }
 
 public enum CoordinateMode
